Adicionado a Autenticação do Admin e Arrumado a validação dos Formularios

+++ routes/web.php
@@ -1,7 +1,11 @@
 <?php
 
 use App\Http\Controllers\LeretecController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
+use Illuminate\Support\Facades\Auth;
+
+
 
 /*
 |--------------------------------------------------------------------------
@@ -20,7 +24,7 @@ Route::get('/', function () {
 
 Route::get('/form', function () {
     return view('leretec/form');
-});
+})->middleware('admin');
 
 Route::get('/login', function () {
     return view('leretec/login');
@@ -36,22 +40,28 @@ Route::get('/historia', function () {
 
 
 
-Route::post('/form', [LeretecController::class,'store']);
+Route::post('/form/cadastrar', [LeretecController::class,'store'])->middleware('admin');
 
 Route::get('/', [LeretecController::class, 'index']);
 
-Route::get('/admin', [LeretecController::class, 'index_admin']);
+Route::get('/admin', [LeretecController::class, 'index_admin'])->middleware('admin');
 
 Route::get('/historia/{id}', [LeretecController::class, 'index_historia']); 
 
-Route::delete('/admin/{id}', [LeretecController::class, 'destroy']);
+Route::delete('/admin/{id}', [LeretecController::class, 'destroy'])->middleware('admin');
 
-Route::get('/admin/edit/{id}', [LeretecController::class, 'edit']);
+Route::get('/admin/edit/{id}', [LeretecController::class, 'edit'])->middleware('admin');
 
-Route::put('/admin/update/{id}', [LeretecController::class, 'update']);
+Route::put('/admin/update/{id}', [LeretecController::class, 'update'])->middleware('admin');
 
 Route::get('/historia/export_user_pdf/{id}', [LeretecController::class, 'export_user_pdf']);
 
 Route::get('/historia/view_user_pdf/{id}', [LeretecController::class, 'view_user_pdf']);
 
 Route::get('/search', [LeretecController::class, 'search']);
+
+
+//admin routes
+Route::post('/auth', [UserController::class, 'auth']);
+
+Route::get('/logout', [UserController::class, 'logout'])->name('logout');
